fixed tearDowns and rewrite `payLoan` to `takeLoan` (#428)
Solved not one... but TWO `FIXME` 's in the tests. The cause was a bug in the situations in which "loans" which turned out to actually be active auctions had the `loans.update` method called on them. This resulted in live auctions being turned back into active loans :scream: 

This PR fixes `tearDown` for two tests:
* `testMergeOrRemoveCollateral`
* `testSettleOnAuctionKicked72HoursAgoAndPartiallyTaken`

Changes:
`loans.update` now accepts an additional argument `inAuction_` which blocks an auction from be added or removed from the loan heap


Challenges:
Not a big fan of how I implemented the change as I don't think its elegant. Would love feeback to figure out a cleaner way to perform this check or better yet remove the check alltogether and refactor how / when / what loans `loans.update` is called on.

Here is another admittedly unelegant idea I had for where we want to call `loans.update`:
       if (Auctions.isActive(auctions, borrowerAddress_)) {
            loans.borrowers[borrowerAddress_] = borrower_;
        } else {
            loans.update(borrowerAddress_, borrower_, loanId);
        }
        

Co-authored-by: Ian Harvey <iharvey@comcast.net>
Co-authored-by: grandizzy <grandizzy.the.egg@gmail.com>
Co-authored-by: grandizzy <38490174+grandizzy@users.noreply.github.com>